+++ core/src/main/java/org/tiqr/core/identity/IdentityDetailFragment.kt
@@ -35,12 +35,12 @@ import androidx.annotation.LayoutRes
 import androidx.hilt.navigation.fragment.hiltNavGraphViewModels
 import androidx.navigation.fragment.findNavController
 import androidx.navigation.fragment.navArgs
+import biometricUsable
 import com.google.android.material.dialog.MaterialAlertDialogBuilder
 import dagger.hilt.android.AndroidEntryPoint
 import org.tiqr.core.R
 import org.tiqr.core.base.BaseFragment
 import org.tiqr.core.databinding.FragmentIdentityDetailBinding
-import org.tiqr.core.util.extensions.biometricUsable
 import org.tiqr.data.model.Identity
 import org.tiqr.data.viewmodel.IdentityViewModel
