+++ gradle/libs.versions.toml
@@ -10,10 +10,10 @@ coroutines = "1.7.3"
 ksp-plugin = "1.9.20-1.0.14"
 
 hilt = "2.48.1"
-navigation = "2.7.5"
+navigation = "2.7.6"
 room = "2.6.1"
 lifecycle = "2.6.2"
-camera = "1.3.0"
+camera = "1.3.1"
 
 okhttp = "4.12.0"
 retrofit = "2.9.0"
@@ -32,7 +32,7 @@ dagger-hilt-compiler = { module = "com.google.dagger:hilt-android-compiler", ver
 dagger-hilt-testing = { module = "com.google.dagger:hilt-android-testing", version.ref = "hilt" }
 dagger-hilt-fragment = "androidx.hilt:hilt-navigation-fragment:1.1.0"
 
-androidx-activity = "androidx.activity:activity-ktx:1.8.1"
+androidx-activity = "androidx.activity:activity-ktx:1.8.2"
 androidx-autofill = "androidx.autofill:autofill:1.1.0"
 androidx-biometric = "androidx.biometric:biometric-ktx:1.2.0-alpha05"
 androidx-concurrent = "androidx.concurrent:concurrent-futures-ktx:1.1.0"
@@ -46,7 +46,7 @@ androidx-testing-core = "androidx.test:core-ktx:1.5.0"
 androidx-testing-epsresso = "androidx.test.espresso:espresso-core:3.5.1"
 androidx-testing-junit = "androidx.test.ext:junit-ktx:1.1.5"
 androidx-testing-rules = "androidx.test:rules:1.5.0"
-androidx-testing-uiautomator = "androidx.test.uiautomator:uiautomator:2.3.0-alpha05"
+androidx-testing-uiautomator = "androidx.test.uiautomator:uiautomator:2.3.0-beta01"
 
 androidx-lifecycle-common = { module = "androidx.lifecycle:lifecycle-common-java8", version.ref = "lifecycle" }
 androidx-lifecycle-livedata = { module = "androidx.lifecycle:lifecycle-livedata-ktx", version.ref = "lifecycle" }
@@ -66,9 +66,9 @@ androidx-camera-camera2 = { module = "androidx.camera:camera-camera2", version.r
 androidx-camera-lifecycle = { module = "androidx.camera:camera-lifecycle", version.ref = "camera" }
 androidx-camera-view = { module = "androidx.camera:camera-view", version.ref = "camera" }
 
-google-android-material = "com.google.android.material:material:1.10.0"
+google-android-material = "com.google.android.material:material:1.11.0"
 google-mlkit-barcode = "com.google.mlkit:barcode-scanning:17.2.0"
-google-firebase-messaging = "com.google.firebase:firebase-messaging:23.3.1"
+google-firebase-messaging = "com.google.firebase:firebase-messaging:23.4.0"
 
 okhttp-okhttp = { module = "com.squareup.okhttp3:okhttp", version.ref = "okhttp" }
 okhttp-logging = { module = "com.squareup.okhttp3:logging-interceptor", version.ref = "okhttp" }
